requirements.txt
flask
Flask-CORS
flask-mysql
PyMySQL